Add "New" option in new shot dialog (#24)
This option allow to create a new scene on the fly (with only a basic camera).

It's probably best practice to create a new scene from a template to keep same scene settings and environment. But user might still want to create a new scene from scratch, without having to do it in a separate step first.